Move planners to their independent submodule
