+++ src/index/plan/index_update_planner.rs
@@ -8,7 +8,8 @@ use crate::{
         predicate::{Expression, Predicate},
         statement::{CreateCommand, UpdateCommand},
     },
-    plan::{select_plan::SelectPlan, table_plan::TablePlan, Plan, PlanControl, UpdatePlanner},
+    plan::{select_plan::SelectPlan, table_plan::TablePlan, Plan, PlanControl},
+    planner::UpdatePlanner,
     record::{field::Value, schema::Schema},
     scan::{RecordId, Scan, ScanControl},
     tx::transaction::Transaction,
